Add unit tests for new code.
