+++ tests/SourcemapToolkit.CallstackDeminifier.UnitTests/SourcemapToolkit.CallstackDeminifier.UnitTests.csproj
@@ -95,6 +95,7 @@
     <Compile Include="FunctionMapConsumerUnitTests.cs" />
     <Compile Include="FunctionMapGeneratorUnitTests.cs" />
     <Compile Include="KeyValueCacheUnitTests.cs" />
+    <Compile Include="SourceMapExtensionsUnitTests.cs" />
     <Compile Include="StackFrameDeminifierUnitTests.cs" />
     <Compile Include="StackTraceDeminifierClosureEndToEndTests.cs" />
     <Compile Include="StackTraceDeminifierMapOnlyEndToEndTests.cs" />
