+++ tests/SourcemapToolkit.SourcemapParser.UnitTests/SourcemapToolkit.SourcemapParser.UnitTests.csproj
@@ -91,6 +91,7 @@
     <Compile Include="Base64ConverterUnitTests.cs" />
     <Compile Include="Base64VlqDecoderUnitTests.cs" />
     <Compile Include="Base64VlqEncoderUnitTests.cs" />
+    <Compile Include="IReadOnlyListExtensionsUnitTests.cs" />
     <Compile Include="NumericMappingEntryUnitTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="MappingsListParserUnitTests.cs" />
